+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/ButtonTypeTest.php
@@ -51,8 +51,8 @@ class ButtonTypeTest extends BaseTypeTestCase
             ->getForm()
             ->createView();
         $this->assertArrayNotHasKey('form', $view->vars['attr']);
-        $this->assertSame($view->vars['id'], $view['child1']->vars['attr']['form']);
-        $this->assertSame($view->vars['id'], $view['child2']->vars['attr']['form']);
+        $this->assertSame($view->vars['form_id'], $view['child1']->vars['attr']['form']);
+        $this->assertSame($view->vars['form_id'], $view['child2']->vars['attr']['form']);
     }
 
     public function testFormAttrOnChild()
@@ -66,7 +66,7 @@ class ButtonTypeTest extends BaseTypeTestCase
             ->getForm()
             ->createView();
         $this->assertArrayNotHasKey('form', $view->vars['attr']);
-        $this->assertSame($view->vars['id'], $view['child1']->vars['attr']['form']);
+        $this->assertSame($view->vars['form_id'], $view['child1']->vars['attr']['form']);
         $this->assertArrayNotHasKey('form', $view['child2']->vars['attr']);
     }
 
@@ -97,7 +97,7 @@ class ButtonTypeTest extends BaseTypeTestCase
             ->createView();
         $this->assertArrayNotHasKey('form', $view->vars['attr']);
         $this->assertSame($stringId, $view->vars['id']);
-        $this->assertSame($view->vars['id'], $view['child1']->vars['attr']['form']);
-        $this->assertSame($view->vars['id'], $view['child2']->vars['attr']['form']);
+        $this->assertSame($view->vars['form_id'], $view['child1']->vars['attr']['form']);
+        $this->assertSame($view->vars['form_id'], $view['child2']->vars['attr']['form']);
     }
 }